remove conf.go move to general conf
